Move valid contribution proof [skip ci]
