+++ types/spectest/tests/partialsigmessage/valid_contribution_proof_metadata.go
@@ -1,4 +1,4 @@
-package messages
+package partialsigmessage
 
 import (
 	"github.com/bloxapp/ssv-spec/qbft"
